+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Extended the import functionality by tags
+- Introduced fuzzy search for the accounts endpoint
 - Improved the language localization for Polish (`pl`)
 - Improved the language localization for Spanish (`es`)